% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayCorrector.R
\name{correct}
\alias{correct}
\title{Correct the bias present in the assay, previously detected by the \code{detect} method}
\usage{
correct(assay,method=NULL,alpha=0.01,type="PA")
}
\arguments{
\item{assay}{The assay to be corrected. Has to be an \code{assay} object.}

\item{method}{\code{NULL}:autodetect (default), \code{1}:additive, \code{2}:multiplicative}

\item{alpha}{Significance level threshold (defaults to 0.05)}

\item{type}{\code{P}:plate-specific, \code{A}:assay-specific, \code{PA}:plate then assay-specific, \code{AP}:assay then plate-specific}
}
\value{
The corrected assay (\code{assay} object)
}
\description{
\code{correct} uses either the additive and multiplicative PMP methods to correct the assay measurements
}

