% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayCorrector.R
\name{create}
\alias{create}
\title{Create a new \code{assay}}
\usage{
assay=create(m)
}
\arguments{
\item{m}{The assay you want to be corrected}

\item{ctrl}{An optional boolean array of the same dimensions as \code{m}. Each entry is 1 if the well is a control well, 0 otherwise. All control wells are excluded from all computations}
}
\description{
\code{create} makes a new object of class assay. You should pass this object to detect() and correct() methods
}
\examples{
m=array(20,c(8,12,5)) # Create array of 20s, 8 rows, 12 columns, 5 plates
m[1:5,12,]=40 # Introduce spatial bias in last column
m=m+rnorm(8*12*5) # Add random additive noise
assay=create(m)
}

