% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayCorrector.R
\name{detect}
\alias{detect}
\title{Detect the type of bias present in the assay}
\usage{
detect(assay,method=1,alpha=0.01,type="P")
}
\arguments{
\item{assay}{The assay to be corrected. Has to be an \code{assay} object.}

\item{alpha}{Significance level threshold (defaults to 0.05)}

\item{type}{\code{P}:plate-specific, \code{A}:assay-specific, \code{PA}:plate then assay-specific, \code{AP}:assay then plate-specific}
}
\value{
The corrected assay (\code{assay} object)
}
\description{
\code{detect} implements both the additive and multiplicative PMP methods
}

