% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m}
\alias{m}
\title{Artificially generated HTS assay}
\format{An object of class \code{array} of dimension 8 x 12 x 5.}
\usage{
m
}
\description{
A 8x12x5 assay (class array) that was generated with partial column bias in the last column.
Gaussian noise was then added
}
\keyword{datasets}

