% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayCorrector.R
\name{plot.assay}
\alias{plot.assay}
\title{Plot assay plate-wise}
\usage{
plot(assay,plate=2,type="P")
}
\arguments{
\item{assay}{The assay you want to plot}

\item{plate}{The plate number}

\item{type}{Either "R" - raw assay, "C" - corrected assay (if it exists) or "P" - spatial bias position}
}
\description{
\code{plot.assay} plots a hit map of the assay (only one plate at a time)
}

