% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plate7}
\alias{plate7}
\title{Plate 7 of raw measurements obtained by Carralot et al. 2012 (Institut Pasteur Korea)}
\format{An object of class \code{array} of dimension 16 x 20 x 1.}
\usage{
plate7
}
\description{
The seventh plate take from an experimental assay, which showed the evidence of existence of multiplicative
kind of bias in screening technologies.
}
\keyword{datasets}
