% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayCorrector.R
\name{plot.assay}
\alias{plot.assay}
\title{Plot assay plate-wise}
\usage{
\method{plot}{assay}(x, ..., plate = 1, type = "R")
}
\arguments{
\item{x}{The assay you want to plot}

\item{...}{Ellipsis to be passed to the \code{levelplotplot()} function from the \code{lattice} package}

\item{plate}{The plate number (Default:1)}

\item{type}{Either "R" - raw assay, "C" - corrected assay (if it exists) or "P" - spatial bias position}
}
\value{
None
}
\description{
\code{plot.assay} plots a hit map of the assay (only one plate at a time)
}
