% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayCorrector.R
\name{print.assay}
\alias{print.assay}
\title{Print assay summary}
\usage{
print(assay,plate=2)
}
\arguments{
\item{assay}{The assay you want to print}

\item{plate}{The plate number}
}
\description{
\code{print.assay} simply prints a summary of the HTS assay
}

